+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/debug/renderer/guice/support/privt/InnerModule.java
@@ -1,6 +1,8 @@
 package ru.vyarus.dropwizard.guice.debug.renderer.guice.support.privt;
 
+import com.google.inject.Exposed;
 import com.google.inject.PrivateModule;
+import com.google.inject.Provides;
 
 /**
  * @author Vyacheslav Rusakov
@@ -14,9 +16,22 @@ public class InnerModule extends PrivateModule {
         bind(InnerService.class);
         bind(OuterService.class);
 
+        bind(OService.class).to(IndirectOuterService.class);
+
         expose(OuterService.class);
+        expose(OService.class);
+    }
+
+    @Provides @Exposed
+    public OuterProviderService getService() {
+        return new OuterProviderService();
     }
 
     public static class InnerService {}
     public static class OuterService {}
+
+    public static class OuterProviderService {}
+
+    public interface OService {}
+    public static class IndirectOuterService implements OService {}
 }